chore: rename from templatecopy to chanelhutt
