+++ lesson_25/chanelhutt/src/components/footer/Footer.tsx
