+++ lesson_25/chanelhutt/src/components/footer/index.ts
