+++ lesson_25/chanelhutt/src/main.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import {RouterProvider, createBrowserRouter} from 'react-router-dom';
 import './index.scss';
-import AddProgram from './pages/AddProgram.tsx';
+import AddProgram from './pages/AddProgram/AddProgram.tsx';
 
 const queryClient = new QueryClient();
