+++ lesson_25/chanelhutt/src/pages/Home/Home.scss
@@ -78,7 +78,7 @@ article {
 .add-program-button {
   display: inline-block;
   padding: 10px;
-  background-color: #007bff;
+  background-color: #f47d26;
   color: #fff;
   font-weight: 600;
   border-radius: 10px;
@@ -86,6 +86,7 @@ article {
   box-shadow: 0 2px 6px rgba(0, 0, 0, 0.2);
   width: 15%;
   transition: background-color 0.3s ease;
+  margin: 1rem;
 }
 
 @media screen and (max-width: 1080px) {
