+++ lesson_25/chanelhutt/src/pages/Home/hero.jpg
