--- lesson_25/templatecopy/src/components/Program/ProgramData.ts
@@ -1,24 +1,0 @@
-export type ProgramType = {
-  title: string;
-  description: string;
-};
-
-export const programs: ProgramType[] = [
-  {
-    title: 'Swine Short Loin',
-    description:
-      'Swine short loin burgdoggen ball tip, shank ham hock bacon...',
-  },
-  {
-    title: 'Bacon Ipsum',
-    description: 'Bacon ipsum dolor amet leberkas chuck biltong pork loin...',
-  },
-  {
-    title: 'Picanha Swine Jowl',
-    description: 'Picanha swine jowl meatball boudin pastrami bresaola...',
-  },
-  {
-    title: 'Kevin Chicken T-Bone',
-    description: 'Kevin chicken t-bone spare ribs shankle bacon drumstick...',
-  },
-];
